# Conflicts:
#	src/main/java/net/swimmingtuna/lotm/events/ClientEventsEntity.java
#	src/main/java/net/swimmingtuna/lotm/events/ModEvents.java

+++ src/main/java/net/swimmingtuna/lotm/item/BeyonderAbilities/Sailor/Roar.java
@@ -5,15 +5,14 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Rarity;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.Vec3;
-import net.swimmingtuna.lotm.caps.BeyonderHolder;
-import net.swimmingtuna.lotm.caps.BeyonderHolderAttacher;
 import net.swimmingtuna.lotm.entity.RoarEntity;
 import net.swimmingtuna.lotm.init.BeyonderClassInit;
 import net.swimmingtuna.lotm.init.EntityInit;
@@ -32,7 +31,7 @@ public class Roar extends SimpleAbilityItem {
     }
 
     @Override
-    public InteractionResult useAbility(Level level, Player player, InteractionHand hand) {
+    public InteractionResult useAbility(Level level, LivingEntity player, InteractionHand hand) {
         if (!checkAll(player)) {
             return InteractionResult.FAIL;
         }
@@ -42,9 +41,8 @@ public class Roar extends SimpleAbilityItem {
         return InteractionResult.SUCCESS;
     }
 
-    public static void roar(Player player) {
+    public static void roar(LivingEntity player) {
         if (!player.level().isClientSide()) {
-            BeyonderHolder holder = BeyonderHolderAttacher.getHolderUnwrap(player);
             RoarEntity roarEntity = new RoarEntity(EntityInit.ROAR_ENTITY.get(), player.level());
             roarEntity.teleportTo(player.getX(), player.getY(), player.getZ());
             Vec3 lookVec = player.getLookAngle();
@@ -58,7 +56,7 @@ public class Roar extends SimpleAbilityItem {
                     forEach(pos -> {
                         if (isInCone(startPos, lookVec, new Vec3(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5), 0.5) && startPos.distanceTo(new Vec3(pos.getX(), pos.getY(), pos.getZ())) <= 10) {
                             BlockState state = player.level().getBlockState(pos);
-                            if (!state.isAir() && state.getDestroySpeed(player.level(), pos) >= 0) {
+                            if (!state.isAir() && state.getDestroySpeed(player.level(), pos) >= 0 && state.getBlock() != Blocks.BEDROCK) {
                                 player.level().destroyBlock(pos, false);
                             }
                         }
@@ -84,6 +82,7 @@ public class Roar extends SimpleAbilityItem {
         tooltipComponents.add(SimpleAbilityItem.getClassText(this.requiredSequence, this.requiredClass.get()));
         super.baseHoverText(stack, level, tooltipComponents, tooltipFlag);
     }
+
     @Override
     public Rarity getRarity(ItemStack pStack) {
         return Rarity.create("SAILOR_ABILITY", ChatFormatting.BLUE);
